Move client into it's own package
