+++ pkg/download/modes.go
@@ -1,9 +1,11 @@
 package download
 
+import "github.com/replicate/pget/pkg/client"
+
 type modeFactoryFunc func() Mode
 
 var modes = map[string]modeFactoryFunc{
-	"buffer":      func() Mode { return &BufferMode{Client: newClient()} },
+	"buffer":      func() Mode { return &BufferMode{Client: client.NewClient()} },
 	"tar-extract": func() Mode { return &ExtractTarMode{} },
 }
