+++ pkg/download/tar_extraction.go
@@ -2,6 +2,7 @@ package download
 
 import (
 	"fmt"
+	"github.com/replicate/pget/pkg/client"
 
 	"github.com/replicate/pget/pkg/extract"
 )
@@ -10,7 +11,7 @@ type ExtractTarMode struct {
 }
 
 func (m *ExtractTarMode) DownloadFile(url string, dest string) error {
-	downloader := &BufferMode{Client: newClient()}
+	downloader := &BufferMode{Client: client.NewClient()}
 	buffer, fileSize, err := downloader.fileToBuffer(url)
 	if err != nil {
 		return fmt.Errorf("error downloading file: %w", err)
